Fixes Performance of landing and hydration errors
